refactor: extract context from `src/api/react.ts`
This is so that this file can be imported by RSC.
