+++ .changeset/thirty-seahorses-roll.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+Extract context from `src/api/react.ts` so that it can be imported in RSC.
