+++ packages/runtime/src/components/page/Page.tsx
@@ -8,7 +8,7 @@ import Head from 'next/head'
 import { BodySnippet } from './BodySnippet'
 import { DocumentReference } from '../../runtimes/react'
 import { createDocumentReference } from '../../state/react-page'
-import { useMakeswiftHostApiClient } from '../../api/react'
+import { useMakeswiftHostApiClient } from '../../next/context/makeswift-host-api-client'
 import { useIsInBuilder } from '../../react'
 import deepEqual from '../../utils/deepEqual'
 import { MakeswiftPageDocument } from '../../next'
